requirements.txt
Django
psycopg2
Pillow
pytest-django
faker
django-disqus
faker
django-hosts
pytils
django-ckeditor
requests